requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
openai==1.3.0
google-cloud-vision==3.4.4
psutil==5.9.6